requirements.txt
django
python-twitter
python-decouple
pydantic
